+++ src/scriptlets/prevent-bab.js
@@ -98,7 +98,13 @@ export function preventBab(source) {
 
 export const preventBabNames = [
     'prevent-bab',
-    // there are no aliases for this scriptlet
+    // aliases are needed for matching the related scriptlet converted into our syntax
+    'ubo-nobab',
+    'nobab',
+    'bab-defuser',
+    'nobab.js',
+    'ubo-nobab.js',
+    'bab-defuser.js',
 ];
 
 // eslint-disable-next-line prefer-destructuring